App rename and backwards compatibility
- App renamed to Moravec (in capital letters)
- Compatibility package name and version code with Moravec 1.0
