+++ android/app/BUCK
@@ -45,12 +45,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.moravec",
+    package = "tedxperiments.math.entrenamente",
 )
 
 android_resource(
     name = "res",
-    package = "com.moravec",
+    package = "tedxperiments.math.entrenamente",
     res = "src/main/res",
 )
