+++ android/app/build.gradle
@@ -98,11 +98,11 @@ android {
     buildToolsVersion "23.0.1"
 
     defaultConfig {
-        applicationId "com.moravec"
+        applicationId "tedxperiments.math.entrenamente"
         minSdkVersion 16
         targetSdkVersion 22
-        versionCode 1
-        versionName "0.0.1"
+        versionCode 22
+        versionName "2.0.0"
         ndk {
             abiFilters "armeabi-v7a", "x86"
         }
