+++ android/app/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.moravec"
-    android:versionCode="1"
-    android:versionName="1.0">
+    package="tedxperiments.math.entrenamente"
+    android:versionCode="22"
+    android:versionName="2.0.0">
 
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.SYSTEM_ALERT_WINDOW"/>
