+++ android/settings.gradle
@@ -1,4 +1,4 @@
-rootProject.name = 'moravec'
+rootProject.name = 'Moravec'
 include ':react-native-svg'
 project(':react-native-svg').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-svg/android')
 include ':react-native-video'
