+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "moravec",
-  "displayName": "moravec"
+  "name": "Moravec",
+  "displayName": "Moravec"
 }
