+++ index.js
@@ -58,4 +58,4 @@ const App = isTest ? CavyMoravecWrapper : Moravec;
 
 export const CDNBaseURL = "https://1841111903.rsc.cdn77.org";
 
-AppRegistry.registerComponent('moravec', () => App);
+AppRegistry.registerComponent('Moravec', () => App);
