+++ ios/Moravec/Images.xcassets/AppIcon.appiconset/1024x1024.png
