+++ ios/Moravec/Images.xcassets/AppIcon.appiconset/120x120-1.png
