+++ ios/Moravec/Images.xcassets/AppIcon.appiconset/40x40.png
