+++ ios/Moravec/Images.xcassets/AppIcon.appiconset/58x58.png
