+++ ios/Moravec/Images.xcassets/AppIcon.appiconset/60x60.png
