+++ ios/Moravec/Images.xcassets/AppIcon.appiconset/80x80.png
