+++ ios/Moravec/Images.xcassets/AppIcon.appiconset/87x87.png
