+++ ios/Moravec/Images.xcassets/AppIcon.appiconset/Contents.json
