+++ ios/MoravecTests/Info.plist
