+++ ios/MoravecTests/MoravecTests.m
@@ -16,11 +16,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React Native!"
 
-@interface moravecTests : XCTestCase
+@interface MoravecTests : XCTestCase
 
 @end
 
-@implementation moravecTests
+@implementation MoravecTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
