+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "moravec",
+  "name": "Moravec",
   "version": "0.0.1",
   "private": true,
   "scripts": {
